requirements.txt
cffi==1.15.0
cycler==0.11.0
fonttools==4.33.3
joblib==1.1.0
kiwisolver==1.4.2
llvmlite==0.38.1
matplotlib==3.5.2
MulticoreTSNE==0.1
numba==0.55.1
numpy==1.21.6
packaging==21.3
Pillow==9.1.1
pycparser==2.21
pynndescent==0.5.7
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.64.0
umap-learn==0.5.3
